Kraya
This is a terminal based application implemented using python. Concepts of dictionaries were used. Graphical representation of sales is included to keep a better track of profitable items.  Option to add a new item to inventory was provided. Two interfaces were introduced for customer and shopkeeper to reduce the confusion. All the items available in shop would be displayed upon request by typing "help".Most sold item and the items which are out of stock would be displayed for future planning in buying items.